+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 
 * Updated Syncfusion controls to v21.2.10
 * Added custom db logging for CUD (CRUD - Read) operations
-* Added LoggingPage to view logs in a ListView
+* Added LoggingPage to view and filter logs in a ListView
 * Moved Constants.cs to Helpers namespace and added Utilities.cs for mutables
 * Updated Settings namespace structure
 * Added BulkEditingPage
@@ -30,8 +30,7 @@
 * Converted static services classes to interfaces
 * Implemented dependencies for services and viewmodels
 * Converted all pngs to svgs
-* Overhauled UI
-* Refactored/cleaned up logic
+* Refactored/cleaned up UI and logic
 * Split up settings page
 
 # MoMoney v1.0.3